Refactor import paths for ScarfPlot and TransitionMatrix components to enhance organization
This commit updates the import paths for the ScarfPlot and TransitionMatrix components, moving them to a more structured utils directory. Additionally, new index files have been created for the plots and shared components, improving the overall organization of the codebase. Unused files related to ScarfPlot state management have been removed, streamlining the project structure and enhancing maintainability.
