+++ src/lib/components/Modal/ModalContent/ModalContentDownloadScarfPlot.svelte
@@ -4,7 +4,7 @@
   import ScarfPlotFigure from '$lib/plots/scarf/components/ScarfPlotFigure.svelte'
   import type { ScarfFillingType } from '$lib/type/Filling/ScarfFilling/ScarfFillingType'
   import GeneralCanvasPreview from '$lib/components/General/GeneralCanvasPreview/GeneralCanvasPreview.svelte'
-  import { transformDataToScarfPlot } from '$lib/plots/scarf/utils/scarfPlotTransformations'
+  import { transformDataToScarfPlot } from '$lib/plots/scarf/utils/transformations'
   import { getParticipants } from '$lib/stores/dataStore'
   import { untrack } from 'svelte'
   import SectionHeader from '$lib/components/Modal/Shared/SectionHeader.svelte'
