+++ src/lib/plots/bar/index.ts
@@ -1,0 +1,8 @@
+// Export bar plot components
+export * from './components'
+
+// Export bar plot types
+export type * from './types'
+
+// Export bar plot utilities
+export * from './utils'
