+++ src/lib/plots/index.ts
@@ -1,0 +1,7 @@
+// Export shared components and utilities
+export * from './shared'
+
+// Export visualization types
+export * from './transition-matrix'
+export * from './scarf'
+export * from './bar'
