+++ src/lib/plots/scarf/components/index.ts
@@ -1,0 +1,11 @@
+// Export scarf plot components
+export { default as ScarfPlot } from './ScarfPlot.svelte'
+export { default as ScarfPlotFigure } from './ScarfPlotFigure.svelte'
+export { default as ScarfPlotButtonMenu } from './ScarfPlotButtonMenu.svelte'
+export { default as ScarfPlotHeader } from './ScarfPlotHeader.svelte'
+export { default as ScarfPlotSelectTimeline } from './ScarfPlotSelectTimeline.svelte'
+export { default as ScarfPlotSelectStimulus } from './ScarfPlotSelectStimulus.svelte'
+export { default as ScarfPlotSelectGroup } from './ScarfPlotSelectGroup.svelte'
+export { default as ScarfPlotButtonZoomIn } from './ScarfPlotButtonZoomIn.svelte'
+export { default as ScarfPlotButtonZoomOut } from './ScarfPlotButtonZoomOut.svelte'
+export { default as ScarfPlotButtonResetView } from './ScarfPlotButtonResetView.svelte'
