+++ src/lib/plots/scarf/index.ts
@@ -1,0 +1,8 @@
+// Export scarf plot components
+export * from './components'
+
+// Export scarf plot types
+export type * from './types'
+
+// Export scarf plot utilities
+export * from './utils'
