+++ src/lib/plots/scarf/types/AoiVisibilityScarfFilling/AoiVisibilityScarfFillingType.ts
