+++ src/lib/plots/scarf/types/AoiVisibilityScarfFilling/SingleAoiVisibilityScarfFillingType.ts
