+++ src/lib/plots/scarf/types/ParticipantScarfFilling/ParticipantScarfFillingType.ts
