+++ src/lib/plots/scarf/types/ScarfFillingType.ts
@@ -1,4 +1,4 @@
-import { AdaptiveTimeline } from '../../../plots/shared/class/AdaptiveTimeline'
+import { AdaptiveTimeline } from '../../shared/class/AdaptiveTimeline'
 import type { StimulusScarfFillingType } from './StimulusScarfFilling/StimulusScarfFillingType'
 import type { StylingScarfFillingType } from './StylingScarfFilling/StylingScarfFillingType'
 import type { ParticipantScarfFillingType } from './ParticipantScarfFilling/ParticipantScarfFillingType'
