+++ src/lib/plots/scarf/types/SegmentScarfFilling/SegmentScarfFillingType.ts
