+++ src/lib/plots/scarf/types/SegmentScarfFilling/SingleSegmentScarfFillingType.ts
