+++ src/lib/plots/scarf/types/StimulusScarfFilling/StimulusScarfFillingType.ts
