+++ src/lib/plots/scarf/types/StylingScarfFilling/SingleStylingScarfFillingType.d.ts
