+++ src/lib/plots/scarf/types/StylingScarfFilling/SingleStylingScarfFillingType.ts
