+++ src/lib/plots/scarf/types/StylingScarfFilling/StylingScarfFillingType.d.ts
