+++ src/lib/plots/scarf/types/StylingScarfFilling/StylingScarfFillingType.ts
