+++ src/lib/plots/shared/class/index.ts
@@ -1,0 +1,2 @@
+// Export shared classes
+export { AdaptiveTimeline } from './AdaptiveTimeline'
