+++ src/lib/plots/shared/types/index.ts
@@ -1,0 +1,7 @@
+/**
+ * Interface for plot dimensions in pixels
+ */
+export interface PlotDimensions {
+  width: number
+  height: number
+}
