+++ src/lib/plots/transition-matrix/index.ts
@@ -1,0 +1,8 @@
+// Export transition matrix components
+export * from './components'
+
+// Export transition matrix types
+export type * from './types'
+
+// Export transition matrix utilities
+export * from './utils'
