--- src/lib/type/PlotState/ScarfPlotState/ScarfPlotStateType.ts
@@ -1,5 +1,0 @@
-export interface ScarfPlotStateType {
-  scarfPlotId: number
-  stimulusId: number
-  zoomLevel: number // 0 = 100%, 1 = 200%, 2 = 400%, 3 = 800%, 4 = 1600%
-}
